Articles from Suragch
I read in the contribution guidelines that I need to make a new request for each article. Since I made quite a few changes I'm going to go ahead and submit this as a single commit. If that doesn't meet your standards I'll understand if you reject the pull request.